backend copy/requirements.txt
arrow==0.14.7
bcrypt==3.2.0
boto3==1.16.47
botocore==1.19.47
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
python-docx==0.8.10
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.25.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
icalendar==4.0.7
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
pycparser==2.20
PyPDF4==1.27.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.22
TatSu==5.5.0
urllib3==1.26.2
Werkzeug==1.0.1
